LayoutTests/inspector/css/nodeLayoutContextTypeChanged-expected.txt

Tests for the CSS.nodeLayoutContextTypeChanged event.



== Running test suite: CSS.nodeLayoutContextTypeChanged
-- Running test case: CSS.nodeLayoutContextTypeChanged.GridToNonGrid
PASS: Layout context should be `grid`.
PASS: Layout context should now be `null`.

-- Running test case: CSS.nodeLayoutContextTypeChanged.NonGridToGrid
PASS: Layout context should be `null`.
PASS: Layout context should now be `grid`.

-- Running test case: CSS.nodeLayoutContextTypeChanged.FlexToNonFlex
PASS: Layout context should be `flex`.
PASS: Layout context should now be `null`.

-- Running test case: CSS.nodeLayoutContextTypeChanged.NonFlexToFlex
PASS: Layout context should be `null`.
PASS: Layout context should now be `flex`.

-- Running test case: CSS.nodeLayoutContextTypeChanged.GridToFlex
PASS: Layout context should now be `grid`.
PASS: Layout context should now be `flex`.

-- Running test case: CSS.nodeLayoutContextTypeChanged.NotFlex.SubmitInput
PASS: Layout context should be `null`.

-- Running test case: CSS.nodeLayoutContextTypeChanged.NotFlex.Select
PASS: Layout context should be `null`.

-- Running test case: CSS.nodeLayoutContextTypeChanged.NotFlex.Button
PASS: Layout context should be `null`.

-- Running test setup.
-- Running test case: CSS.nodeLayoutContextTypeChanged.NotFlex.DynamicallyAddedButton
PASS: Layout context should be `null`.
